+++ zuihou-commons/zuihou-boot/src/main/java/com/github/zuihou/boot/handler/GlobalDbExceptionHandler.java
