--- zuihou-commons/zuihou-boot/src/main/resources/j2cache/network.xml
@@ -1,43 +1,0 @@
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xmlns="urn:org:jgroups"
-        xsi:schemaLocation="urn:org:jgroups http://www.jgroups.org/schema/jgroups.xsd">
-    <UDP
-            mcast_port="${jgroups.udp.mcast_port:45588}"
-            ip_ttl="4"
-            ucast_recv_buf_size="5M"
-            ucast_send_buf_size="5M"
-            mcast_recv_buf_size="5M"
-            mcast_send_buf_size="5M"
-            max_bundle_size="64K"
-            enable_diagnostics="false"
-
-            thread_pool.min_threads="2"
-            thread_pool.max_threads="8"
-            thread_pool.keep_alive_time="5000"/>
-
-    <PING/>
-    <MERGE3 max_interval="30000"
-            min_interval="10000"/>
-    <FD_SOCK/>
-    <FD_ALL/>
-    <VERIFY_SUSPECT timeout="1500"/>
-    <pbcast.NAKACK2 xmit_interval="500"
-                    xmit_table_num_rows="100"
-                    xmit_table_msgs_per_row="2000"
-                    xmit_table_max_compaction_time="30000"
-                    use_mcast_xmit="false"
-                    discard_delivered_msgs="true"/>
-    <UNICAST3 xmit_interval="500"
-              xmit_table_num_rows="100"
-              xmit_table_msgs_per_row="2000"
-              xmit_table_max_compaction_time="60000"
-              conn_expiry_timeout="0"/>
-    <pbcast.STABLE desired_avg_gossip="50000"
-                   max_bytes="4M"/>
-    <pbcast.GMS print_local_addr="true" join_timeout="2000"/>
-    <UFC max_credits="2M"
-         min_threshold="0.4"/>
-    <MFC max_credits="2M"
-         min_threshold="0.4"/>
-    <FRAG2 frag_size="60K"/>
-</config>
