+++ zuihou-commons/zuihou-cloud-starter/src/main/java/com/github/zuihou/cloud/http/OkHttpSlf4jLogger.java
@@ -1,4 +1,4 @@
-package com.github.zuihou.openfeign.http;
+package com.github.zuihou.cloud.http;
 
 import lombok.extern.slf4j.Slf4j;
